douyin_ecom_analyzer/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
requests>=2.25.0
tqdm>=4.60.0
streamlit>=1.10.0
plotly>=5.0.0
pyyaml>=6.0 